fix: 불필요한 파일 삭제
