+++ RF/Services/SignUp/EmailService.swift
@@ -49,7 +49,7 @@ final class EmailService {
                        parameters: body,
                        encoder: JSONParameterEncoder.default)
             .validate(statusCode: 200..<201)
-            .responseDecodable(of: SignUpBase.self) { response in
+            .responseDecodable(of: ResponseData.self) { response in
                 switch response.result{
                 case .success(let data):
                     observer.onNext(data.isSuccess ?? false)
